\name{vs30}
\alias{vs30}
\title{Calculate Vs(30)}
\description{
Simple function for quickly calculating the average shear wave velocity
in the upper 30 meters from the output of Rvelslant().
}

\usage{
vs30(mod)
}
\arguments{
  \item{mod}{a list, as returned by \code{Rvelslant}.}

}
\value{the average shear-wave velocity in the upper 30 meters.
}
\author{Eric M. Thompson <eric.thompson@tufts.edu>}
\references{
  
}
\examples{


}
\keyword{}
